front end for other entities
