+++ frontend/src/pages/ActivityReport/Pages/components/__tests__/OtherEntity.js
@@ -5,24 +5,20 @@ import { render, screen, waitFor } from '@testing-library/react';
 import React from 'react';
 import fetchMock from 'fetch-mock';
 import { FormProvider, useForm } from 'react-hook-form/dist/index.ie11';
-import ObjectivePicker from '../ObjectivePicker';
+import OtherEntity from '../OtherEntity';
 
 // eslint-disable-next-line react/prop-types
-const RenderObjective = ({ objectivesWithoutGoals }) => {
+const RenderOtherEntity = ({ objectivesWithoutGoals }) => {
   const hookForm = useForm({
     mode: 'onChange',
     defaultValues: {
       objectivesWithoutGoals,
-      collaborators: [],
-      author: {
-        role: 'Central Office',
-      },
     },
   });
 
   return (
     <FormProvider {...hookForm}>
-      <ObjectivePicker />
+      <OtherEntity roles={['Central Office']} />
     </FormProvider>
   );
 };
@@ -46,23 +42,23 @@ const objectives = [
   },
 ];
 
-describe('ObjectivePicker', () => {
+describe('OtherEntity', () => {
   beforeEach(async () => {
     fetchMock.restore();
     fetchMock.get('/api/topic', []);
   });
   it('renders created objectives', async () => {
-    render(<RenderObjective objectivesWithoutGoals={objectives} />);
+    render(<RenderOtherEntity objectivesWithoutGoals={objectives} />);
 
     const title = await screen.findByText('title');
     expect(title).toBeVisible();
   });
 
   it('the button adds a new objective', async () => {
-    render(<RenderObjective objectivesWithoutGoals={[]} />);
+    render(<RenderOtherEntity objectivesWithoutGoals={[]} />);
     const button = await screen.findByRole('button', { name: /Add new objective/i });
-    expect(screen.queryByText(/objective status/i)).toBeNull();
+    expect(screen.queryAllByText(/objective status/i).length).toBe(1);
     userEvent.click(button);
-    await waitFor(() => expect(screen.queryByText(/objective status/i)).not.toBeNull());
+    await waitFor(() => expect(screen.queryAllByText(/objective status/i).length).toBe(2));
   });
 });
